requirements.txt
google-cloud-firestore
streamlit